fix: be less strict with negative moments of inertia (#690)
This allows setting different thresholds for almost-zero moments of inertia, updates the default threshold, and adds a (previously failing) passing real-world example.

It also corrects a parameter type in the docs, removes some white space, and fixes ruff call arguments in CI

---------

+++ overreact/coords.py
@@ -2011,7 +2011,7 @@ def _equivalent_atoms(
         Atomic coordinates.
     method : str, optional
         Method of partitioning: "atommass" (same atoms same groups), "cluster".
-    thresh : int, optional
+    thresh : float, optional
         Threshold to consider atom clusters.
 
     Returns
